requirements.txt
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
xgboost==1.7.6
lightgbm==4.0.0
joblib==1.3.2
pyyaml==6.0.1
jupyter==1.0.0
plotly==5.15.0
scipy==1.11.1
